completely changed poll()
 - `Servblr.poll` is now like most of the other methods, it receives a chat_id
as the first argument and operates only on it.
 - `poll` is shortcut-ed on the `Charblr` class
 - We got rid of that fekker of wait_time dict. We have instead two
phases, a normal phase and a sleep phase. The keyword argument
`sleep_threshold` defines when to enter the sleep phase. In the normal
phase, we keep getting the actual messages (using Servblr.get_message)
and see if there are new among them. But in the sleep mode we poll first
for unread counts using `Servblr.get_counts` which uses the endpoint
/counts, which is how Tumblr website gets its updates. Only after we know
there are new messages that we get them as in the normal phase.
 - Loggers are not not flooded anymore with specifically the sleep handler
and the poll results. Now we only log when the phase changes and when the
results are not empty they are not empty.
 - A part of code also has a ton load of useless comments that are not
of any good for anything. But they say you should comment your code. I
commented a part of it heavily.